pyproject.toml
[tool.poetry]
name = "limnc-flaked"
version = "0.1.0"
description = "LéXPLORE datalakes (service)"
authors = ["ymarcon <yannick.marcon@epfl.ch>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.scripts]
flaked = "limnc_flaked.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
apscheduler = "^3.11.0"
fastapi = "^0.115.8"
pysftp = "^0.2.9"
pyyaml = "^6.0.2"
typer = "^0.15.1"
uvicorn = "^0.34.0"
tenacity = "^9.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
